Separate the command handling from the actual fetch logic so we can use it in another tool directly
